pyproject.toml
[project]
name = "react-agent-with-langgraph"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.8",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
